yolo-instance/requirements.txt
opencv-python==4.5.3.56
torch==1.8.0
torchvision==0.9.0
natsort
ffmpeg-python
pytorchvideo
scipy
pandas
requests
pillow